backend/src/requirements.txt
aiohttp
colorama
databricks-sql-connector
defog==0.64.2
fastapi
google-auth
google-cloud-bigquery
google-cloud-storage
hypercorn
matplotlib
mysql-connector-python
numpy
openai
pandas
pandasql
pgvector
pika
psycopg2-binary
pyyaml
redshift_connector
requests
scipy
seaborn
slack_sdk
snowflake-connector-python
sqlalchemy
statsmodels
tabulate
tiktoken
torch